+++ src/main/java/mate/academy/internetshop/model/Item.java
@@ -8,9 +8,9 @@ public class Item {
     private String name;
     private BigDecimal price;
 
-    public Item(String name, double price) {
+    public Item(String name, BigDecimal price) {
         this.name = name;
-        this.price = BigDecimal.valueOf(price);
+        this.price = price;
     }
 
     public Long getItemId() {
